add runtime model caching
