+++ core/tools/model_loader.py
@@ -9,44 +9,66 @@ runtime_configuration = get_runtime_config()
 llm_config = runtime_configuration.llm_config
 embedder_config = runtime_configuration.embedder_config
 
-# todo: for this to be memory efficient,
-#       we have to cache returns as singletons, and serve them
-# problem with the caching requests above: we have to share those singletons across instances
+# problem with the current caching: we have to share those singletons across instances
 
 # fixme: n_gpu_layers=-1 is a poor approach, it can and will cause crashes.
 #        with llama.cpp we have to manually calculate and set this number
 
 
-def load_ollama_llm():
-    return Ollama(model=llm_config.model_name)
+def load_ollama_llm() -> Ollama:
+    cached_llm = runtime_configuration.llm_object
+    if cached_llm:
+        return cached_llm
+    else:
+        llm = Ollama(model=llm_config.model_name)
+        runtime_configuration.llm_object = llm
+        return llm
 
 
-def load_ollama_embedder():
-    return OllamaEmbeddings(model=embedder_config.model_name)
+def load_ollama_embedder() -> OllamaEmbeddings:
+    cached_embedder = runtime_configuration.embedder_object
+    if cached_embedder:
+        return cached_embedder
+    else:
+        embedder = OllamaEmbeddings(model=embedder_config.model_name)
+        runtime_configuration.embedder_object = embedder
+        return embedder
 
 
 def load_hf_llm():
-    base_model_path = hf_hub_download(
-        llm_config.model_file, filename=llm_config.model_name
-    )
-    return Llama(
-        model_path=base_model_path,
-        n_gpu_layers=-1,
-        n_batch=llm_config.model_token_limit,
-        verbose=True,
-    )
+    cached_llm = runtime_configuration.llm_object
+    if cached_llm:
+        return cached_llm
+    else:
+        base_model_path = hf_hub_download(
+            llm_config.model_file, filename=llm_config.model_name
+        )
+        llm = Llama(
+            model_path=base_model_path,
+            n_gpu_layers=-1,
+            n_batch=llm_config.model_token_limit,
+            verbose=True,
+        )
+        runtime_configuration.llm_object = llm
+        return llm
 
 
 def load_hf_embedder():
-    embedder_model_path = hf_hub_download(
-        embedder_config.model_file, filename=embedder_config.model_name
-    )
-    return Llama(
-        model_path=embedder_model_path,
-        n_gpu_layers=-1,
-        n_batch=embedder_config.model_token_limit,
-        verbose=True,
-    )
+    cached_embedder = runtime_configuration.embedder_object
+    if cached_embedder:
+        return cached_embedder
+    else:
+        embedder_model_path = hf_hub_download(
+            embedder_config.model_file, filename=embedder_config.model_name
+        )
+        embedder = Llama(
+            model_path=embedder_model_path,
+            n_gpu_layers=-1,
+            n_batch=embedder_config.model_token_limit,
+            verbose=True,
+        )
+        runtime_configuration.embedder_object = embedder
+        return embedder
 
 
 def load_llm():
